requirements.txt
ccxt
requests
websocket-client==0.58.0

// pip install -r requirements.txt